+++ lib/src/models/completion_request.dart
@@ -1,4 +1,5 @@
 import 'package:chat_gpt_flutter/src/models/chat_gpt_model.dart';
+import 'package:chat_gpt_flutter/src/models/message.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'completion_request.g.dart';
@@ -6,21 +7,23 @@ part 'completion_request.g.dart';
 @JsonSerializable(fieldRename: FieldRename.snake)
 class CompletionRequest {
   final ChatGptModel model;
-  final List<String>? prompt;
-  final int? maxTokens;
+  final List<Message> messages;
   final double? temperature;
   final double? topP;
   final int? n;
   final bool? stream;
+  final String? stop;
+  final int? maxTokens; // max_tokens + messages tokens > 4096 will throw 400 error code
 
   CompletionRequest({
-    this.model = ChatGptModel.textDavinci003,
-    required this.prompt,
+    this.model = ChatGptModel.gpt35Turbo,
+    required this.messages,
     this.temperature = 0,
-    this.maxTokens = 16,
     this.topP,
-    this.n = 1,
+    this.n,
     this.stream,
+    this.stop,
+    this.maxTokens = 16,
   }) : assert(!(temperature != null && topP != null));
 
   factory CompletionRequest.fromJson(Map<String, dynamic> data) =>