+++ lib/src/models/delta_response.g.dart
@@ -1,0 +1,17 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'delta_response.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+DeltaResponse _$DeltaResponseFromJson(Map<String, dynamic> json) =>
+    DeltaResponse(
+      content: json['content'] as String? ?? '',
+    );
+
+Map<String, dynamic> _$DeltaResponseToJson(DeltaResponse instance) =>
+    <String, dynamic>{
+      'content': instance.content,
+    };