+++ lib/src/models/message.dart
@@ -1,0 +1,19 @@
+import 'package:json_annotation/json_annotation.dart';
+
+part 'message.g.dart';
+
+@JsonSerializable(fieldRename: FieldRename.snake)
+class Message {
+  final String role;
+  final String content;
+
+  Message({
+    required this.role,
+    required this.content,
+  });
+
+  factory Message.fromJson(Map<String, dynamic> data) =>
+      _$MessageFromJson(data);
+
+  Map<String, dynamic> toJson() => _$MessageToJson(this);
+}